Move namespace dependent information to access ticket.
Topic paths are dependent upon the namespace and environment the agent is connecting to.
To avoid having the agent to construct these paths itself, the paths are instead returned in the access ticket.
The agent now only has to connect to the frontdoorservice in the correct namespace and environment in order to get the correct topic paths.